minor reformating
