+++ warp10/src/main/java/io/warp10/script/formatted/RunAndGenerateDocumentationWithUnitTests.java
@@ -16,8 +16,10 @@
 
 package io.warp10.script.formatted;
 
+import io.warp10.WarpConfig;
 import io.warp10.script.*;
 import io.warp10.script.formatted.FormattedWarpScriptFunction.Arguments;
+import org.junit.BeforeClass;
 
 import java.io.*;
 import java.lang.reflect.Method;
@@ -89,6 +91,14 @@ public class RunAndGenerateDocumentationWithUnitTests {
   //
   // Tests
   //
+
+  @BeforeClass
+  final public static void beforeClass() throws Exception {
+    StringBuilder props = new StringBuilder();
+
+    props.append("warp.timeunits=us");
+    WarpConfig.safeSetProperties(new StringReader(props.toString()));
+  }
 
   final protected void generate(List<String> functionNames) throws Exception {
